Refacto datasets TODO, update the unit tests Signed-off-by: arthurPignet <arthur.pignet@mines-paristech.fr>
